requirements.txt
# Core
langchain==1.2.0
langchain-chroma==1.1.0
langchain-openai==1.1.6
chromadb==1.3.7

# LLM & Embeddings
groq==0.37.1
openai==1.109.1

# Data processing
python-frontmatter==1.1.0
pyyaml
python-dotenv==1.2.1

# Search
rank-bm25==0.2.2
numpy==2.2.6

# UI
gradio==6.2.0

# Evaluation
ragas==0.4.2
datasets==4.4.1

# Utils
requests==2.32.5
huggingface-hub==0.36.0